pytest test was failing because of lstm pytorch type code missing (#2016)
quickly fix to make "make test" work. it tests all the code and some
tests were failing because of a recent merge. this fixes that

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211011174894293)

+++ agent/src/metta/agent/lib/lstm.py
@@ -88,7 +88,7 @@ class LSTM(LayerBase):
 
         hidden = rearrange(hidden, "(b t) h -> t b h", b=B, t=TT)
 
-        training_env_id_start = td.get("training_env_id_start", torch.tensor([0]))[0].item()
+        training_env_id_start = td.get("training_env_id_start", torch.tensor([0], dtype=torch.long))[0].item()
 
         if training_env_id_start in self.lstm_h and training_env_id_start in self.lstm_c:
             h_0 = self.lstm_h[training_env_id_start]
